+++ tests/integration/test_use_cases/test_10_asset_upload.py
@@ -13,9 +13,8 @@ from starfish.asset import DataAsset
 
 def test_10_asset_upload(resources, remote_agent):
     test_data = secrets.token_hex(1024)
-    asset = DataAsset.create('TestAsset', test_data)
-    listing = remote_agent.register_asset(asset, resources.listing_data)
-    assert(listing)
-    assert(listing.asset)
+    asset_data = DataAsset.create('TestAsset', test_data)
+    asset = remote_agent.register_asset(asset_data)
+    assert(asset)
 
-    assert(remote_agent.upload_asset(listing.asset))
+    assert(remote_agent.upload_asset(asset))