+++ tests/integration/test_use_cases/test_13_asset_listing_search.py
@@ -13,7 +13,9 @@ from starfish.asset import DataAsset
 
 def test_13_asset_listing_search(resources, remote_agent):
     test_data = secrets.token_hex(1024)
-    asset = DataAsset.create('TestAsset', test_data)
-    listing = remote_agent.register_asset(asset, resources.listing_data)
+    asset_data = DataAsset.create('TestAsset', test_data)
+    asset = remote_agent.register_asset(asset_data)
+    assert(asset)
+    listing = remote_agent.create_listing(resources.listing_data, asset.did)
     assert(listing)
-    assert(listing.asset)
+    assert(listing.asset_did)