+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 ## Change Log
 
-### Release TBD
+### Release v0.7.0
+
+    **Warning** Breaking API changes
 
 +   better linting and module import order
 +   provenance
++   Split up register_asset and create_listing in agents
++   Changed squid to use register_asset_and_listing
++   change listing to pass the asset_did instead of the asset object
 
 ### Release v0.6.5